requirements.txt
openai>=1.3.0
langchain>=0.0.350
langchain-openai>=0.0.2
faiss-cpu>=1.7.4
pypdf>=3.17.0
tiktoken>=0.5.1
python-dotenv>=1.0.0
streamlit>=1.28.0
